Add Exception responses

+++ Server/src/main/java/com/example/server/service/ItemService.java
@@ -1,5 +1,7 @@
 package com.example.server.service;
 
+import com.example.server.exception.ItemException.ErrorCode;
+import com.example.server.exception.ItemException.ItemException;
 import com.example.server.mapper.ItemMapper;
 import com.example.server.model.Item;
 import com.example.server.model.ItemEnum.ItemStatus;
@@ -16,33 +18,43 @@ public class ItemService {
     private ItemMapper itemMapper;
 
     @Autowired
+    private UserService userService;
+
+    @Autowired
     private StringRedisTemplate redisTemplate;
 
-    public List<Item> selectAllItems(){
+    public List<Item> selectAllItems() {
         return itemMapper.selectAllItems();
     }
 
-    public Item selectItemById(Integer id){
+    public Item selectItemById(Integer id) {
         return itemMapper.selectItemById(id);
     }
 
-    public Item insertItem(Item item){
-        itemMapper.insertItem(item.getId(), item.getSellerId(),item.getName(), item.getPrice(), item.getTag().toString(), item.getStatus().toString(), item.getDescription());
-        return item;
+    public void insertItem(Item item) {
+        if (selectItemById(item.getId()) != null) {
+            throw new ItemException(ErrorCode.DUPLICATE_ITEM, "Item with this id already exists!");
+        }
+        if (userService.selectUserById(item.getSellerId()) == null) {
+            throw new ItemException(ErrorCode.NO_EXISTING_SELLER, "The seller does not exist!");
+        }
+
+        itemMapper.insertItem(item.getId(), item.getSellerId(), item.getName(), item.getPrice(), item.getTag().toString(), item.getStatus().toString(), item.getDescription());
     }
 
-    public Item deleteItemById(Integer id){
-        Item itemDeleted=selectItemById(id);
+    public void deleteItemById(Integer id) {
+        if (selectItemById(id) == null) {
+            throw new ItemException(ErrorCode.NO_EXISTING_ITEM, "Item with this id already exists!");
+        }
         itemMapper.deleteItemById(id);
-        return itemDeleted;
     }
-
-
 
 
     @Transactional
-    public Item updateItemStatus(Integer id, ItemStatus status){
-        itemMapper.updateItemStatus(id,status.toString());
-        return selectItemById(id);
+    public void updateItemStatus(Integer id, ItemStatus status) {
+        if (selectItemById(id) == null) {
+            throw new ItemException(ErrorCode.NO_EXISTING_ITEM, "Item with this id already exists!");
+        }
+        itemMapper.updateItemStatus(id, status.toString());
     }
 }
